chore(chartbeat): move sdk loader to new file
