+++ src/integrations/Chartbeat/browser.js
@@ -6,7 +6,7 @@ import {
   INTEGRATION_LOAD_CHECK_INTERVAL,
 } from '../../utils/constants';
 import { NAME } from './constants';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
+import { loader } from './loader';
 
 class Chartbeat {
   constructor(config, analytics, destinationInfo) {
@@ -108,16 +108,7 @@ class Chartbeat {
   initAfterPage() {
     onBody(() => {
       const script = this.isVideo ? 'chartbeat_video.js' : 'chartbeat.js';
-      function loadChartbeat() {
-        const e = document.createElement('script');
-        const n = document.getElementsByTagName('script')[0];
-        e.type = 'text/javascript';
-        e.async = true;
-        e.src = `//static.chartbeat.com/js/${script}`;
-        e.setAttribute('data-loader', LOAD_ORIGIN);
-        n.parentNode.insertBefore(e, n);
-      }
-      loadChartbeat();
+      loader(script);
     });
 
     this._isReady(this).then((instance) => {
